updates to feat eng task notebook and predict notebook
