+++ .databricks/.databricks.env
@@ -1,6 +1,6 @@
 DATABRICKS_AUTH_TYPE=metadata-service
 DATABRICKS_HOST=https://e2-demo-field-eng.cloud.databricks.com/
-DATABRICKS_METADATA_SERVICE_URL=http://127.0.0.1:49429/615128ff-4675-4f3d-8404-360a9b31bd24
+DATABRICKS_METADATA_SERVICE_URL=http://127.0.0.1:49429/2bef0de0-dc7a-48bf-9cdf-a894b1db0248
 DATABRICKS_PROJECT_ROOT=/Users/logan.rupert/Library/CloudStorage/GoogleDrive-logan.rupert@databricks.com/My Drive/Repositories/NHL_SOG
 PYDEVD_WARN_SLOW_RESOLVE_TIMEOUT=10
 SPARK_CONNECT_USER_AGENT=databricks-vscode/1.4.0
