--- nhlPredict/src/preprocess_model/requirements.txt
@@ -1,2 +1,0 @@
-mlflow==2.11.3
-cloudpickle==2.2.1
